ref: move processAnnualReview into processListsBeforeAndDuringStart
