+++ src/scheduler/workers/processListsBeforeAndDuringStart/helpers.ts
